Fix for Mac Certs
From my work it does not appear that there is any programmatic way to recover
the public/private key pair in a Key Chain without having to enter a password.
As a result, we are going to grab the certificates directly off the disk, like
we do on Linux machines.

+++ src/tools/CertHelper/Program.cs
@@ -1,6 +1,7 @@
 ﻿using Azure.Identity;
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Specialized;
+using System.Runtime.InteropServices;
 using System.Security.Cryptography.X509Certificates;
 using System.Text;
 
@@ -8,52 +9,55 @@ namespace CertHelper;
 
 internal class Program
 {
+
     static readonly string TENANT_ID = "72f988bf-86f1-41af-91ab-2d7cd011db47";
     static readonly string CERT_CLIENT_ID = "8c4b65ef-5a73-4d5a-a298-962d4a4ef7bc";
+
     static async Task<int> Main(string[] args)
     {
-        try
+        if (!RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
         {
-            var kvc = new KeyVaultCert();
-            await kvc.LoadKeyVaultCertsAsync();
-            if (kvc.ShouldRotateCerts())
+            Console.WriteLine("Not Here");
+            try
             {
-                using (var localMachineCerts = new X509Store(StoreName.My, StoreLocation.CurrentUser))
+                var kvc = new KeyVaultCert();
+                await kvc.LoadKeyVaultCertsAsync();
+                if (kvc.ShouldRotateCerts())
                 {
-                    localMachineCerts.Open(OpenFlags.ReadWrite);
-                    localMachineCerts.RemoveRange(kvc.LocalCerts.Certificates);
-                    localMachineCerts.AddRange(kvc.KeyVaultCertificates);
+                    using (var localMachineCerts = new X509Store(StoreName.My, StoreLocation.CurrentUser))
+                    {
+                        localMachineCerts.Open(OpenFlags.ReadWrite);
+                        localMachineCerts.RemoveRange(kvc.LocalCerts.Certificates);
+                        localMachineCerts.AddRange(kvc.KeyVaultCertificates);
+                    }
+                }
+                var bcc = new BlobContainerClient(new Uri("https://pvscmdupload.blob.core.windows.net/certstatus"),
+                    new ClientCertificateCredential(TENANT_ID, CERT_CLIENT_ID, kvc.KeyVaultCertificates.First()));
+                var currentKeyValutCertThumbprints = "";
+                foreach (var cert in kvc.KeyVaultCertificates)
+                {
+                    currentKeyValutCertThumbprints += $"[{DateTimeOffset.UtcNow}] {cert.Thumbprint}{Environment.NewLine}";
+                }
+                var blob = bcc.GetBlobClient(System.Environment.MachineName);
+                if (blob.Exists())
+                {
+                    var result = blob.DownloadContent();
+                    var currentBlob = result.Value.Content.ToString();
+                    currentBlob = currentBlob + currentKeyValutCertThumbprints;
+                    blob.Upload(new MemoryStream(Encoding.UTF8.GetBytes(currentBlob)), overwrite: true);
+                }
+                else
+                {
+                    blob.Upload(new MemoryStream(Encoding.UTF8.GetBytes(currentKeyValutCertThumbprints)), overwrite: false);
                 }
             }
-            var bcc = new BlobContainerClient(new Uri("https://pvscmdupload.blob.core.windows.net/certstatus"),
-                new ClientCertificateCredential(TENANT_ID, CERT_CLIENT_ID, kvc.KeyVaultCertificates.First()));
-            var currentKeyValutCertThumbprints = "";
-            foreach(var cert in kvc.KeyVaultCertificates)
+            catch (Exception ex)
             {
-                currentKeyValutCertThumbprints += $"[{DateTimeOffset.UtcNow}] {cert.Thumbprint}{Environment.NewLine}";
+                Console.WriteLine("Failed to rotate certificates");
+                Console.WriteLine(ex.Message);
+                Console.WriteLine(ex.StackTrace);
             }
-            var blob = bcc.GetBlobClient(System.Environment.MachineName);
-            if (blob.Exists())
-            {
-                var result = blob.DownloadContent();
-                var currentBlob = result.Value.Content.ToString();
-                currentBlob = currentBlob + currentKeyValutCertThumbprints;
-                blob.Upload(new MemoryStream(Encoding.UTF8.GetBytes(currentBlob)), overwrite: true);
-            }
-            else
-            {
-                blob.Upload(new MemoryStream(Encoding.UTF8.GetBytes(currentKeyValutCertThumbprints)), overwrite: false);
-            }
-            
         }
-        catch (Exception ex)
-        {
-            Console.WriteLine("Failed to rotate certificates");
-            Console.WriteLine(ex.Message);
-            Console.WriteLine(ex.StackTrace);
-        }
-
-
 
         using (var store = new X509Store(StoreName.My, StoreLocation.CurrentUser, OpenFlags.ReadWrite))
         {
